% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{adaptive_memory}
\alias{adaptive_memory}
\title{Apply adaptive memory}
\usage{
adaptive_memory(search_history, memory)
}
\arguments{
\item{search_history}{a list of prior solutions.}

\item{memory}{the maximum length of search_history}
}
\value{
a list of prior solutions.
}
\description{
Apply an adaptive memory to a list of prior solutions.
}
\details{
Certain heuristics, like tabu search, have an adaptive memory that maintains a
search history of a specified length. This function prunes a list of prior solutions
to a specified length.
}
\examples{
adaptive_memory(list("010", "210", "100", "120"), 3)

}
