% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{binary_max}
\alias{binary_max}
\title{Calculate maximum feasible value of binary}
\usage{
binary_max(binary_sol)
}
\arguments{
\item{binary_sol}{a list of Boolean values corresponding to a binary string.}
}
\value{
a number.
}
\description{
Get maximum possible value for based on the length of a binary.
}
\details{
A 2-bit binary vector will have a maximum value of 3. A 3-bit binary vector
will have a maximum value of 7.
}
\examples{
binary_max(list("1", "1", "0", "0", "1"))

}
