% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_solution_representation.R
\name{binary_round}
\alias{binary_round}
\title{Binary Round}
\usage{
binary_round(init_sol, neighbors, tabu = list())
}
\arguments{
\item{init_sol}{A string representing the initial solution.}

\item{neighbors}{A number specifying the number of neighbors to iterate through.}

\item{tabu}{(Optional) A tabu list.}
}
\value{
a list containing a list of the updated binary lists, a list of the
updated binary lists converted to strings, a list of the updated binary strings
converted to numbers, and a list of the positions that were flipped.
}
\description{
Generates round of binary strings.
}
\details{
Pending.
}
