% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{binary_to_number}
\alias{binary_to_number}
\title{Convert binary string to number}
\usage{
binary_to_number(binary_sol)
}
\arguments{
\item{binary_sol}{a list of Boolean values corresponding to a binary string.}
}
\value{
a number.
}
\description{
Translates a binary list to a number.
}
\details{
Binary lists are lists of Boolean values that represent a binary string. For example,
the binary list list(F, T, F, T, T) represents the binary vector 01011.

Binary strings are translated to binary lists using 'string_to_binary()'.
}
