% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{binary_to_string}
\alias{binary_to_string}
\title{Convert binary list to string}
\usage{
binary_to_string(binary_sol)
}
\arguments{
\item{binary_sol}{a list of Boolean values corresponding to a binary vector.}
}
\value{
a string.
}
\description{
Translate binary list to string.
}
\details{
Binary lists are lists of Boolean values that represent a binary string. For example,
the binary list list(F, T, F, T, T) represents the binary vector 01011.

This function is used to format algorithm output.
}
