% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{conduct_flow_jobs}
\alias{conduct_flow_jobs}
\title{Job Shop - Flow}
\usage{
conduct_flow_jobs(param_list)
}
\arguments{
\item{param_list}{a list of job shop parameters, including (and in this order) the initial
solution string; job IDs  p (processing time), d (due date),
and w (weight for delays); number of machines.}
}
\value{
a list object that contains job completion information for each job
and the weighted delays
}
\description{
Conducts the entire flow job shop process.
}
\details{
Pending

Does not provide an optimal solution.
}
