% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_toy_data.R
\docType{data}
\name{data_1_mach_n_jobs}
\alias{data_1_mach_n_jobs}
\title{Set Covering Data}
\format{
a matrix.
}
\source{
OR 670 - Lecture 2 Excel examples.
}
\usage{
data_1_mach_n_jobs
}
\description{
Data to use in one machine, n jobs job shop problems.
}
\keyword{datasets}
