% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_toy_data.R
\docType{data}
\name{data_set_covering}
\alias{data_set_covering}
\title{Set Covering Data}
\format{
a matrix.
}
\source{
OR 670 - Lecture 2 Excel examples.
}
\usage{
data_set_covering
}
\description{
Data to use in set covering problems.
}
\keyword{datasets}
