% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_toy_data.R
\docType{data}
\name{data_tsp}
\alias{data_tsp}
\title{Traveling Salesman}
\format{
a matrix.
}
\source{
OR 670 - Lecture 2 Excel examples.
}
\usage{
data_tsp
}
\description{
Data in traveling salesman problems.
}
\keyword{datasets}
