% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_stopping_criteria.R
\name{elapsed_time}
\alias{elapsed_time}
\title{Check elapsed time}
\usage{
elapsed_time(sc_list, stop_crit)
}
\arguments{
\item{sc_list}{A list of algorithm output that can be evaluated by stopping functions.}

\item{stop_crit}{The value to be used when determining if an algorithm should stop.}
}
\value{
a Boolean value.
}
\description{
Checks elapsed time against stopping criteria.
}
\details{
Pending.
}
