% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{exchange_values}
\alias{exchange_values}
\title{Exchange values in a list}
\usage{
exchange_values(sol_list, one, two)
}
\arguments{
\item{sol_list}{a solution list.}

\item{one}{the first position in the solution list that will be exchanged.}

\item{two}{the second position in the solution list that will be exchanged.}
}
\value{
a solution list.
}
\description{
Swap values in solution list.
}
\details{
Pending.
}
