% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{flip_positions}
\alias{flip_positions}
\title{Flip values in binary list}
\usage{
flip_positions(sol_length, neighbors)
}
\arguments{
\item{sol_length}{the length of a solution list.}

\item{neighbors}{the number of positions to flip in the solution list.}
}
\value{
a list of position pairs to flip.
}
\description{
Generate list of positions to flip.
}
\details{
Pending.
}
\examples{
flip_positions(5, 2)
}
