% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_methods.R
\name{genetic_algorithm}
\alias{genetic_algorithm}
\title{Genetic Algorithm}
\usage{
genetic_algorithm(
  genes,
  crossover_probability,
  crossover_points,
  mutation_rate,
  elite_parents = 0,
  initial_gen,
  fitness_function,
  fitness_function_params = NULL,
  selection_function,
  crossover_function,
  mutation_function,
  stop_func,
  stop_crit,
  permutation = F
)
}
\description{
Executes the genetic algorithm.
}
\details{
Pending.
}
