% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{job}
\alias{job}
\title{job}
\usage{
job(job_info)
}
\arguments{
\item{job_info}{a list of job information.}
}
\value{
a list (job object) to be used by 'job_shop'.
}
\description{
Creates a single job for a job shop process.
}
\details{
This function takes in a list of job information that includes p (job length), d (due date), and
w (weight for delays).
}
