% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{job_shop}
\alias{job_shop}
\title{Job Shop Problem}
\usage{
job_shop(param_list)
}
\arguments{
\item{param_list}{a list of job shop parameters, including (and in this order) the initial
solution string; a matrix of job information with column order job ID, p (processing time), d (due date),
and w (weight for delays); number of machines; and job shop type in a string.}
}
\value{
a list object that contains the total process time (machine_time) and the
total weighted delay (total_weighted_delay).
}
\description{
Conducts the entire job shop process, from job object creation to total weighted delay calculations.
}
\details{
Given a parameter list of job shop information, creates job objects and sends them
through the user-specified job shop type function.

Does not provide an optimal solution.
}
