% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_methods.R
\name{local_search_storage}
\alias{local_search_storage}
\title{Local Search (Storage)}
\usage{
local_search_storage(
  obj_func,
  problem_params,
  sol_represent,
  eval_func,
  stop_func,
  initial_solution,
  neighbors,
  stop_crit
)
}
\arguments{
\item{obj_func}{The objective function.}

\item{problem_params}{A list of all parameters required to run the objective function.}

\item{sol_represent}{A function that outputs the required solution representation.}

\item{eval_func}{A function that evaluates the algorithm's output.}

\item{stop_func}{A function that evaluates the heuristic's metadata to determine if the search should end.}

\item{initial_solution}{A string representing the initial solution.}

\item{neighbors}{A number representing the number of neighbors to iterate through.}

\item{stop_crit}{A number used by the stop_func function to determine if the search should stop.}
}
\value{
a table of heuristic output including time to execute each round, round ID, each
round's solution, each round's function output, each previous round's function output, and
the improvement between each round.
}
\description{
Executes the local search algorithm and returns the results of all rounds.
}
\details{
Pending.
}
