% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{m_machines_n_jobs_flow}
\alias{m_machines_n_jobs_flow}
\title{m machines, n jobs (flow)}
\usage{
m_machines_n_jobs_flow(jobs, machines, min_max = 1)
}
\arguments{
\item{jobs}{a list of job objects.}

\item{machines}{the number of machines in the process.}

\item{min_max}{(Optional) Used to specify whether the next-queue decision should be based on
maximum (default) or minimum total queue value.}
}
\value{
a list object that contains the total process time (machine_time) and the
total weighted delay (total_weighted_delay).
}
\description{
Conducts an m machine, n jobs process in which the jobs enter the i+1th machine after
completing in the ith machine.
}
\details{
Requires a list of job objects (i.e. list of lists of job information) generated by
'job()'.
}
