% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{matrix_setup}
\alias{matrix_setup}
\title{Matrix setup}
\usage{
matrix_setup(data, state_names = list(), row_sum_check = T)
}
\arguments{
\item{data}{a list of rbind-ed rows.}

\item{state_names}{(Optional) a list of row and column names list(list(row_names), list(column_names)). If not provided, uses alphabetical labels.}

\item{row_sum_check}{(Optional) Boolean; should the rows sum to one? Default is True.}
}
\value{
a list of prior solutions.
}
\description{
Creates a named, valid transition matrix from input.
}
\details{
Used to set up matrices.
}
