% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_evaluation.R
\name{min_val}
\alias{min_val}
\title{Choose minimum objective function output value}
\usage{
min_val(xs, ys)
}
\arguments{
\item{xs}{A list of objective function solutions (x values) in its raw form (a list of values).}

\item{ys}{A list of objective function outputs (y values).}
}
\value{
a list containing the nest solution  or the yielding the minimum objective function output in its
raw form (a list of values), a cleaned version of the best solution, the solution's value, and the
position of the best solution in the solution list.
}
\description{
Identify the solution that yields the min value from a list of solutions.
}
\details{
Pending.
}
