% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{node_connection_exists}
\alias{node_connection_exists}
\title{Check if node connection exists}
\usage{
node_connection_exists(pair1, pair2)
}
\arguments{
\item{pair1}{the c(row, column) of a matrix cell.}

\item{pair2}{the c(row, column) of a matrix cell.}
}
\value{
a Boolean value.
}
\description{
Checks for connection between two nodes.
}
\details{
Compares row, column information of two nodes to determine connection.
}
\examples{
node_connection_exists(c(1, 3), c(3, 1))

}
