% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{node_paths}
\alias{node_paths}
\title{Generate feasible paths through nodes}
\usage{
node_paths(pairs)
}
\arguments{
\item{pairs}{a list of pairs.}
}
\value{
a matrix.
}
\description{
Returns matrix of connected edges.
}
\details{
Pending.
}
