% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_stopping_criteria.R
\name{number_of_rounds}
\alias{number_of_rounds}
\title{Check number of rounds}
\usage{
number_of_rounds(sc_list, stop_crit)
}
\arguments{
\item{sc_list}{A list of algorithm output that can be evaluated by stopping functions.}

\item{stop_crit}{The value to be used when determining if an algorithm should stop.}
}
\value{
a Boolean value.
}
\description{
Checks number of rounds against stopping criteria.
}
\details{
Pending.
}
