% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{one_machine_n_jobs}
\alias{one_machine_n_jobs}
\title{one machine, n jobs}
\usage{
one_machine_n_jobs(jobs, machines = 1, delay = 0)
}
\arguments{
\item{jobs}{a list of job objects.}

\item{machines}{(Optional) number of machines; default is obviously 1}

\item{delay}{(Optional) used if a delay needs to be added to the whole process.}
}
\value{
a list object that contains the total process time (machine_time) and the
total weighted delay (total_weighted_delay).
}
\description{
Conducts a one machine, n jobs process.
}
\details{
Requires a list of job objects (i.e. list of lists of job information) generated by
'job()'.
}
