% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_solution_representation.R
\name{permutation_round}
\alias{permutation_round}
\title{Permutation Round}
\usage{
permutation_round(init_sol, neighbors, tabu = list())
}
\arguments{
\item{init_sol}{A string representing the initial solution.}

\item{neighbors}{A number specifying the number of neighbors to iterate through.}

\item{tabu}{(Optional) A tabu list.}
}
\value{
a list containing a list of the updated string lists, a list of the
updated lists converted to strings, a list of the updated strings
converted to function input, and a list of the positions that were flipped.
}
\description{
Generates round of permuted strings.
}
\details{
Pending.
}
