% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_stopping_criteria.R
\name{relative_improvement}
\alias{relative_improvement}
\title{Check relative improvement}
\usage{
relative_improvement(sc_list, stop_crit)
}
\arguments{
\item{sc_list}{A list of algorithm output that can be evaluated by stopping functions.}

\item{stop_crit}{The value to be used when determining if an algorithm should stop.}
}
\value{
a Boolean value.
}
\description{
Checks improvement between rounds against stopping criteria.
}
\details{
Pending.
}
