% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{set_covering}
\alias{set_covering}
\title{Set Covering Problem}
\usage{
set_covering(param_list)
}
\arguments{
\item{param_list}{a list of job shop parameters, including (and in this order) the initial
solution string; a matrix of set covering problem information; the maximum node value for
inclusion in the set; the penalty to apply per uncovered node.}
}
\value{
the value of the valid nodes.
}
\description{
Calculates value of a specific solution to a set covering problem.
}
\details{
Creates a list of nodes that meet the minimum value criteria (valid), creates a path
through the valid nodes, then calculates the value of the valid nodes.

Does not provide an optimal solution.
}
