% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{string_list_to_bool}
\alias{string_list_to_bool}
\title{Convert list of strings to list of Boolean}
\usage{
string_list_to_bool(str_list)
}
\arguments{
\item{str_list}{a list of characters representing a binary string.}
}
\value{
a list of Boolean values corresponding to a binary string.
}
\description{
Translate list of 0s and 1s, stored as characters, to a list of Boolean values.
}
\details{
Binary lists  are initially entered as strings and need to be translated to a
list of Boolean values for use by this package. The function 'string_to_list()'
can convert a string to a list of characters, which can then be converted to a
list of Boolean values by this function.
}
\examples{
string_list_to_bool(list("1", "1", "0", "0", "1"))

}
