% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{string_to_bool}
\alias{string_to_bool}
\title{Convert string to Boolean}
\usage{
string_to_bool(str_sol)
}
\arguments{
\item{str_sol}{a string of characters.}
}
\value{
a list of Boolean values corresponding to a binary string.
}
\description{
Translates a solution string to a binary list.
}
\details{
Solution strings can represent sequence information or a binary string.
Binary lists are lists of Boolean values that represent a binary string. For example,
the binary list list(F, T, F, T, T) represents the binary vector 01011.
}
\examples{
string_to_bool("2143")

}
