% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_utilities.R
\name{string_to_list}
\alias{string_to_list}
\title{Convert string to list}
\usage{
string_to_list(str_sol)
}
\arguments{
\item{str_sol}{a string of characters.}
}
\value{
a list of characters.
}
\description{
Translates a solution string to list.
}
\details{
Solution strings represent sequence information.
}
\examples{
string_to_list("2143")
}
