% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_problems.R
\name{traveling_salesman}
\alias{traveling_salesman}
\title{Traveling Sales Problem}
\usage{
traveling_salesman(solution_list, round_trip = T)
}
\arguments{
\item{solution_list}{a string of characters.}

\item{round_trip}{(Optional) Boolean; is the path a round trip or does it end without returning.}
}
\value{
the total distance traveled.
}
\description{
Calculates total distance traveled in traveling salesman problem given specific node order and matrix of edge values.
}
\details{
This function does not find the optimal path.
}
