% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_solution_representation.R
\name{update_binary}
\alias{update_binary}
\title{Update Binary List}
\usage{
update_binary(binary_sol, position)
}
\arguments{
\item{binary_sol}{A binary list, or list of Boolean values.}

\item{position}{A number or vector of numbers specifying the position(s) in the binary list to be flipped.}
}
\value{
a binary list.
}
\description{
Flips a Boolean value in a binary list.
}
\details{
Pending.
}
